typ header is optional
https://openid.net/specs/draft-jones-json-web-token-07.html#ReservedClaimName
"The typ (type) claim is used to declare a type for the contents of this JWT Claims Set. The typ value is case sensitive. This claim is OPTIONAL."

Apple's identity token used for Apple Sign In
currently does not include a typ header. It just has the alg and kid.

ex.
```json
{
  "kid": "YuyXoY",
  "alg": "RS256"
}
```